+++ src/WebLinterTest/EslintTest.cs
@@ -16,7 +16,7 @@ namespace WebLinterTest
         [TestMethod, TestCategory("ESLint")]
         public void Standard()
         {
-            var result = LinterFactory.Lint(Settings.CWD, Settings.Instance, "../../artifacts/eslint/a.js");
+            var result = LinterFactory.Lint(Settings.Instance, "../../artifacts/eslint/a.js");
             var first = result.First();
             Assert.IsTrue(first.HasErrors);
             Assert.IsTrue(first.Errors.First().IsError, "Severity is not 'error'");
@@ -27,7 +27,7 @@ namespace WebLinterTest
         [TestMethod, TestCategory("ESLint")]
         public void MultipleInput()
         {
-            var result = LinterFactory.Lint(Settings.CWD, Settings.Instance, "../../artifacts/eslint/a.js", "../../artifacts/eslint/b.js");
+            var result = LinterFactory.Lint(Settings.Instance, "../../artifacts/eslint/a.js", "../../artifacts/eslint/b.js");
             Assert.IsTrue(result.First().HasErrors);
             Assert.AreEqual(6, result.First().Errors.Count);
         }
@@ -35,16 +35,16 @@ namespace WebLinterTest
         [TestMethod, TestCategory("ESLint")]
         public void JSX()
         {
-            var result = LinterFactory.Lint(Settings.CWD, Settings.Instance, "../../artifacts/eslint/a.jsx");
+            var result = LinterFactory.Lint(Settings.Instance, "../../artifacts/eslint/a.jsx");
             Assert.IsTrue(result.First().HasErrors);
-            Assert.IsTrue(result.First().Errors.First().Message.Contains("react/display-name"), "Unexpected error message");
+            Assert.AreEqual("react/display-name", result.First().Errors.First().ErrorCode, "Unexpected error message");
             Assert.AreEqual(2, result.First().Errors.Count);
         }
 
         [TestMethod, TestCategory("ESLint")]
         public void FileNotExist()
         {
-            var result = LinterFactory.Lint(Settings.CWD, Settings.Instance, "../../artifacts/eslint/doesntexist.js");
+            var result = LinterFactory.Lint(Settings.Instance, "../../artifacts/eslint/doesntexist.js");
             Assert.IsTrue(result.First().HasErrors);
         }
     }